(Caution) Can call a service to replace a short url w/ a custom path. Caution: This is dangerous if deployed with no authentication required! Anyone could invalidate existing short URLs!

+++ Durwella.UrlShortening/UrlShortener.cs
@@ -44,8 +44,11 @@ namespace Durwella.UrlShortening
             var directUrl = UrlUnwrapper.GetDirectUrl(url);
             if (Repository.ContainsKey(customHash))
                 throw new ArgumentException("The given custom short URL is already in use.");
-            var oldKey = Repository.GetKey(directUrl);
-            Repository.Remove(oldKey);
+            if (Repository.ContainsValue(directUrl))
+            {
+                var oldKey = Repository.GetKey(directUrl);
+                Repository.Remove(oldKey);
+            }
             Repository.Add(customHash, directUrl);
             return CompleteUrl(customHash);
         }
